+++ Riot/Managers/Theme/Themes/DarkTheme.swift
@@ -26,7 +26,9 @@ class DarkTheme: NSObject, Theme {
 
     var backgroundColor: UIColor = UIColor(rgb: 0x15191E)
 
-    var baseColor: UIColor = BuildSettings.newAppLayoutEnabled ? UIColor(rgb: 0x15191E) : UIColor(rgb: 0x21262C)
+    var baseColor: UIColor {
+        BuildSettings.isNewAppLayoutActivated ? UIColor(rgb: 0x15191E) : UIColor(rgb: 0x21262C)
+    }
     var baseIconPrimaryColor: UIColor = UIColor(rgb: 0xEDF3FF)
     var baseTextPrimaryColor: UIColor = UIColor(rgb: 0xFFFFFF)
     var baseTextSecondaryColor: UIColor = UIColor(rgb: 0xA9B2BC)
@@ -35,7 +37,9 @@ class DarkTheme: NSObject, Theme {
     var searchPlaceholderColor: UIColor = UIColor(rgb: 0xA9B2BC)
     var searchResultHighlightColor: UIColor = UIColor(rgb: 0xFCC639).withAlphaComponent(0.3)
 
-    var headerBackgroundColor: UIColor = BuildSettings.newAppLayoutEnabled ? UIColor(rgb: 0x15191E) : UIColor(rgb: 0x21262C)
+    var headerBackgroundColor: UIColor {
+        BuildSettings.isNewAppLayoutActivated ? UIColor(rgb: 0x15191E) : UIColor(rgb: 0x21262C)
+    }
     var headerBorderColor: UIColor  = UIColor(rgb: 0x15191E)
     var headerTextPrimaryColor: UIColor = UIColor(rgb: 0xFFFFFF)
     var headerTextSecondaryColor: UIColor = UIColor(rgb: 0xA9B2BC)